move pasword manager logic to api calls
