+++ packages/dappmanager/src/calls/index.ts
@@ -48,8 +48,7 @@ export { packageSetEnvironment } from "./packageSetEnvironment.js";
 export { packageSetPortMappings } from "./packageSetPortMappings.js";
 export { packageStartStop } from "./packageStartStop.js";
 export { packageGettingStartedToggle } from "./packageGettingStartedToggle.js";
-export { passwordChange } from "./passwordChange.js";
-export { passwordIsSecure } from "./passwordIsSecure.js";
+export { passwordChange, passwordIsSecure } from "./passwordManager.js";
 export { poweroffHost } from "./poweroffHost.js";
 export { portsApiStatusGet } from "./portsStatusGet.js";
 export { portsUpnpStatusGet } from "./portsStatusGet.js";
