+++ packages/dappmanager/src/calls/passwordManager.ts
@@ -1,3 +1,4 @@
+import * as db from "@dappnode/db";
 import { shell } from "@dappnode/utils";
 import { getDappmanagerImage } from "@dappnode/dockerapi";
 
@@ -12,7 +13,7 @@ const baseCommand = `docker run --rm -v /etc:/etc --privileged --entrypoint=""`;
  *
  * @returns true = is secure / false = is not
  */
-export async function isPasswordSecure(): Promise<boolean> {
+async function isPasswordSecure(): Promise<boolean> {
   const image = await getDappmanagerImage();
   try {
     const res = await shell(
@@ -36,7 +37,7 @@ export async function isPasswordSecure(): Promise<boolean> {
  *
  * @param newPassword = "super-secure-password"
  */
-export async function changePassword(newPassword: string): Promise<void> {
+async function changePassword(newPassword: string): Promise<void> {
   if (!newPassword) throw Error("newPassword must be defined");
   if (typeof newPassword !== "string")
     throw Error("newPassword must be a string");
@@ -63,4 +64,48 @@ export async function changePassword(newPassword: string): Promise<void> {
   await shell(
     `${baseCommand} -e PASS='${newPassword}' ${image} sh -c 'echo dappnode:$PASS | chpasswd'`
   );
+}
+
+// API calls
+
+/**
+ * Checks if the user `dappnode`'s password in the host machine
+ * is NOT the insecure default set at installation time.
+ * It does so by checking if the current salt is `insecur3`
+ *
+ * - This check will be run every time this node app is started
+ *   - If the password is SECURE it will NOT be run anymore
+ *     and this call will return true always
+ *   - If the password is INSECURE this check will be run every
+ *     time the admin requests it (on page load)
+ *
+ * @returns true = is secure / false = is not
+ */
+export async function passwordIsSecure(): Promise<boolean> {
+  if (db.passwordIsSecure.get()) {
+    return true;
+  } else {
+    const isSecure = await isPasswordSecure();
+    if (isSecure) db.passwordIsSecure.set(isSecure);
+    return isSecure;
+  }
+}
+
+/**
+ * Changes the user `dappnode`'s password in the host machine
+ * Only allows it if the current password has the salt `insecur3`
+ *
+ * @param newPassword super-secure-password
+ */
+export async function passwordChange({
+  newPassword
+}: {
+  newPassword: string;
+}): Promise<void> {
+  if (!newPassword) throw Error("Argument newPassword must be defined");
+
+  await changePassword(newPassword);
+
+  // Update the DB "is-password-secure" check
+  await passwordIsSecure();
 }
