--- packages/dappmanager/src/modules/README.md
@@ -1,1 +1,0 @@
-# Modules
